feat: treat snippets and groupId as pointers
