+++ types.go
@@ -189,66 +189,6 @@ type client struct {
 	debug      io.Writer
 }
 
-// Check type constants
-type CheckType string
-
-// TypeBrowser is used to identify a browser check.
-const TypeBrowser = "BROWSER"
-
-// TypeAPI is used to identify an API check.
-const TypeAPI = "API"
-
-// Escalation type constants
-
-// RunBased identifies a run-based escalation type, for use with an AlertSettings.
-const RunBased = "RUN_BASED"
-
-// TimeBased identifies a time-based escalation type, for use with an AlertSettings.
-const TimeBased = "TIME_BASED"
-
-// Assertion source constants
-
-// StatusCode identifies the HTTP status code as an assertion source.
-const StatusCode = "STATUS_CODE"
-
-// JSONBody identifies the JSON body data as an assertion source.
-const JSONBody = "JSON_BODY"
-
-// TextBody identifies the response body text as an assertion source.
-const TextBody = "TEXT_BODY"
-
-// Headers identifies the HTTP headers as an assertion source.
-const Headers = "HEADERS"
-
-// ResponseTime identifies the response time as an assertion source.
-const ResponseTime = "RESPONSE_TIME"
-
-// Assertion comparison constants
-
-// Equals asserts that the source and target are equal.
-const Equals = "EQUALS"
-
-// NotEquals asserts that the source and target are not equal.
-const NotEquals = "NOT_EQUALS"
-
-// IsEmpty asserts that the source is empty.
-const IsEmpty = "IS_EMPTY"
-
-// NotEmpty asserts that the source is not empty.
-const NotEmpty = "NOT_EMPTY"
-
-// GreaterThan asserts that the source is greater than the target.
-const GreaterThan = "GREATER_THAN"
-
-// LessThan asserts that the source is less than the target.
-const LessThan = "LESS_THAN"
-
-// Contains asserts that the source contains a specified value.
-const Contains = "CONTAINS"
-
-// NotContains asserts that the source does not contain a specified value.
-const NotContains = "NOT_CONTAINS"
-
 // Check represents the parameters for an existing check.
 type Check struct {
 	ID                        string                     `json:"id"`
@@ -267,14 +207,14 @@ type Check struct {
 	DoubleCheck               bool                       `json:"doubleCheck"`
 	Tags                      []string                   `json:"tags,omitempty"`
 	SSLCheck                  bool                       `json:"sslCheck"`
-	SetupSnippetID            int64                      `json:"setupSnippetId,omitempty"`
-	TearDownSnippetID         int64                      `json:"tearDownSnippetId,omitempty"`
-	LocalSetupScript          string                     `json:"localSetupScript,omitempty"`
-	LocalTearDownScript       string                     `json:"localTearDownScript,omitempty"`
+	SetupSnippetID            *int64                     `json:"setupSnippetId"`
+	TearDownSnippetID         *int64                     `json:"tearDownSnippetId"`
+	LocalSetupScript          string                     `json:"localSetupScript"`
+	LocalTearDownScript       string                     `json:"localTearDownScript"`
 	AlertSettings             AlertSettings              `json:"alertSettings,omitempty"`
 	UseGlobalAlertSettings    bool                       `json:"useGlobalAlertSettings"`
 	Request                   Request                    `json:"request"`
-	GroupID                   int64                      `json:"groupId,omitempty"`
+	GroupID                   *int64                     `json:"groupId"`
 	GroupOrder                int                        `json:"groupOrder,omitempty"`
 	AlertChannelSubscriptions []AlertChannelSubscription `json:"alertChannelSubscriptions,omitempty"`
 }
@@ -375,10 +315,10 @@ type Group struct {
 	DoubleCheck               bool                       `json:"doubleCheck"`
 	UseGlobalAlertSettings    bool                       `json:"useGlobalAlertSettings"`
 	AlertSettings             AlertSettings              `json:"alertSettings,omitempty"`
-	SetupSnippetID            int64                      `json:"setupSnippetId,omitempty"`
-	TearDownSnippetID         int64                      `json:"tearDownSnippetId,omitempty"`
-	LocalSetupScript          string                     `json:"localSetupScript,omitempty"`
-	LocalTearDownScript       string                     `json:"localTearDownScript,omitempty"`
+	SetupSnippetID            *int64                     `json:"setupSnippetId"`
+	TearDownSnippetID         *int64                     `json:"tearDownSnippetId"`
+	LocalSetupScript          string                     `json:"localSetupScript"`
+	LocalTearDownScript       string                     `json:"localTearDownScript"`
 	AlertChannelSubscriptions []AlertChannelSubscription `json:"alertChannelSubscriptions,omitempty"`
 }
